```text refactor: Update files table migration to use bigInteger for size column
The migration file '2024_05_21_150417_create_files_table.php' has been modified to change the data type of the 'size' column from integer to bigInteger. This change allows for larger file sizes to be stored accurately in the database.

+++ database/migrations/2024_05_21_150417_create_files_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('basename');
             $table->string('extension');
             $table->string('mime_type')->nullable();
-            $table->integer('size')->nullable();
+            $table->bigInteger('size')->nullable();
             $table->dateTime('date_added')->nullable();
 
             $table->timestamps();
